Check the most recent 14 days of data for new games.
Games are not added to the database in chronological order. For example, a set of games from one region may be uploaded before a set of games from another region, but the second region's games happened before the first. This would make the second regions games to be skipped since every new download only searches after the end of the already downloaded data.

Fix this by removing the most recent 14 games of data, and then re-downloading this data from the database. This will work as long as game results are added to the database less than 14 days after they are played (afaik, this is always the case barring some freak accident).